fix(ios): swizzle UIApplication.sendEvent to avoid breaking keyboard input
